Attempt to clean better WIP files during failover
Files caught in the very middle of the write would not have complete
block for the Backup(Hot standby node).

So LeaseManager.checklease would call FSNamesystem.internalReleaseLease
to receover the lease ( rememenber it set all to expired).

If the file was set as underconstruction or in recovery file would not
be imediately closed. It would try to recever the block => restore the
minimum of replicas. But if the NN failed on the middle of the writting
we cannot garantee that there is the minimum replicas because the write
failed as well. We also expecting clients to rewrite the file. So we
dont care if the blocks have to be recovery or not as they will be
removed when client starting rewriting the file from scratch

So we close all WIP files with no replications checks to keep system in
 consistent state and clear all leases.

That was the idea behind the last commit, but we wanted the system to
it by itself. So our hopes that setting leases to expired, WIP files
would be closed  in a clean way. And clients could rewrite again.
It prove to no be working in some cases so we decide to do things
explicitily